chore: pin ty, fix ty 0.0.42 findings, and drop if TYPE_CHECKING
CI's typecheck installs the latest ty (uv.lock is gitignored, so it re-resolves every run), which surfaced 9 diagnostics the older local ty missed. Pin ty==0.0.42 in the dev specs for determinism and resolve all findings: ColumnKey wrapping in filterable_table, a Select cast in settings_screen, and targeted ty: ignore for type-only (loguru record) and App-subclass attribute access.

Remove all 'if TYPE_CHECKING:' blocks per the new convention: use top-level runtime imports, and relocate the shared user-stat dataclasses into a dependency-free stoei/usage_stats.py to break the formatters -> user_overview import cycle (re-exported from user_overview so existing imports stay unchanged). Documented the convention in AGENTS.md.
