+++ pyproject.toml
@@ -37,7 +37,7 @@ dev = [
     "pytest-asyncio>=0.24.0",
     "pytest-cov>=6.0.0",
     "ruff>=0.8.2",
-    "ty>=0.0.1a6",
+    "ty==0.0.42",
     "pre-commit>=4.0.1",
 ]
 
@@ -113,5 +113,5 @@ dev = [
     "pytest>=9.0.2",
     "pytest-asyncio>=1.3.0",
     "ruff>=0.14.13",
-    "ty>=0.0.12",
+    "ty==0.0.42",
 ]
