+++ stoei/colors.py
@@ -16,11 +16,7 @@ Usage:
 from __future__ import annotations
 
 from dataclasses import dataclass
-from typing import TYPE_CHECKING, Any
-
-if TYPE_CHECKING:
-    pass
-
+from typing import Any
 
 # Fallback colors when theme is not available (using sensible defaults)
 FALLBACK_COLORS = {
